{{cookiecutter.project_name}}/requirements.txt
alembic==1.7.7
amqp==5.1.1
autopep8==1.6.0
billiard==3.6.4.0
black==22.3.0
blinker==1.4
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
flake8==4.0.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sentry-sdk==1.5.8
six==1.16.0
SQLAlchemy==1.4.32
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webargs==8.1.0
Werkzeug==2.0.3
zipp==3.8.0
